Upgrade to work with Heist 0.10
This is just a mindless port to make it build.  The magnitude of the changes
in Heist 0.10 really merit a deeper look to see how this package should be
changed to support compiled splices, but I don't have time for that right now.
I'm also a little unsure about the appropriate version bump.  Since the
external API is essentially unchanged, it seems like my minor version bump
might be ok, but I could also conceive of it causing problems.