Allow dfListItem to be used multiple times
This generalizes things so that your view can have multiple places where it
loops through the list of list items.  One case where this is useful is if
each item form is complex and you want to summarize it in one place on the
page with say a name and have the rest of the item become visible on a
different part of the page when selected.  I've tested this on one complex
form and it seems to work nicely, but it could probably use a bit more
testing.